added mocks for gql queries
