+++ __tests__/Chat.test.tsx
@@ -3,74 +3,11 @@ import { fireEvent, screen, waitFor } from '@testing-library/react-native';
 import customRender from '../utils/jestRender';
 
 import Chat from '../screens/Chat';
-import { GET_CONTACTS } from '../graphql/queries/Contact';
-
-const mockContacts = {
-  id: '1',
-  name: 'test',
-  maskedPhone: '12*****90',
-  lastMessageAt: '2021-08-10T12:00:00.000Z',
-};
-
-const noSearchMocks = [
-  {
-    request: {
-      query: GET_CONTACTS,
-      variables: {
-        filter: { term: '' },
-        messageOpts: { limit: 1 },
-        contactOpts: { limit: 10 },
-      },
-    },
-    result: {
-      data: {
-        search: [
-          {
-            contact: mockContacts,
-            messages: [
-              {
-                id: '1',
-                body: 'test message',
-              },
-            ],
-          },
-        ],
-      },
-    },
-  },
-];
-
-const withSearchMocks = [
-  {
-    request: {
-      query: GET_CONTACTS,
-      variables: {
-        filter: { term: 'test mock' },
-        messageOpts: { limit: 1 },
-        contactOpts: { limit: 10 },
-      },
-    },
-    result: {
-      data: {
-        search: [
-          {
-            contact: mockContacts,
-            messages: [
-              {
-                id: '1',
-                body: 'test message',
-              },
-            ],
-          },
-        ],
-      },
-    },
-  },
-];
+import { NO_SEARCH_CONTACTS_MOCK, SEARCH_CONTACTS_MOCK } from '../__mocks__/queries/contact';
 
 describe('Chat screen', () => {
   test('renders correctly', async () => {
-    const { getByTestId, findByText } = customRender(<Chat />, noSearchMocks);
+    const { getByTestId, findByText } = customRender(<Chat />, NO_SEARCH_CONTACTS_MOCK);
     const searchInput = getByTestId('searchInput');
     const searchIcon = getByTestId('searchIcon');
     const filterIcon = getByTestId('filterIcon');
@@ -91,7 +28,7 @@ describe('Chat screen', () => {
   });
 
   test('updates search correctly', async () => {
-    const { getByTestId } = customRender(<Chat />, withSearchMocks);
+    const { getByTestId } = customRender(<Chat />, SEARCH_CONTACTS_MOCK);
 
     const searchInput = getByTestId('searchInput');
     fireEvent.changeText(searchInput, 'test search');
@@ -107,7 +44,7 @@ describe('Chat screen', () => {
 
     const { getByTestId } = customRender(
       <Chat navigation={{ navigate: navigateMock }} />,
-      withSearchMocks
+      SEARCH_CONTACTS_MOCK
     );
 
     const searchIcon = getByTestId('searchIcon');
@@ -120,7 +57,7 @@ describe('Chat screen', () => {
   });
 
   test('should test menu is visible on press or not ', async () => {
-    const { getByTestId } = customRender(<Chat />, noSearchMocks);
+    const { getByTestId } = customRender(<Chat />, NO_SEARCH_CONTACTS_MOCK);
     await waitFor(() => {
       fireEvent.press(getByTestId('menuIcon'));
     });
